requirements.txt
flask
flask_cors
flask_sqlalchemy
requests
openai
psycopg2-binary
gunicorn
python-dotenv
PyJWT>=2.8.0
